requirements.txt
appdirs==1.4.3
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.10.0
distlib==0.3.0
Django==4.0.3
django-crispy==1.8.1
django-crispy-forms==1.14.0
django-filter==22.1
django-widget-tweaks==1.4.12
docopt==0.6.2
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
idna==3.3
kiwisolver==1.3.1
matplotlib==3.3.3
mccabe==0.6.1
numpy==1.19.4
Pillow==8.0.1
pipreqs==0.4.11
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.6
python-gettext==4.1
requests==2.27.1
rosetta==0.3
six==1.14.0
sorl-thumbnail==12.9.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
virtualenv==20.0.17
yarg==0.1.9
